Merge feature/details-screen to develop

+++ GalleryApp/GalleryApp/Application/AppDelegate.swift
@@ -12,10 +12,9 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        
         let cache = ImageCache.default
         cache.memoryStorage.config.totalCostLimit = 10 * 1024 * 1024
-        cache.diskStorage.config.sizeLimit = 100 * 1024 * 1024
+        cache.diskStorage.config.sizeLimit = UInt(100 * 1024 * 1024)
 
         return true
     }
